requirements.txt
flask
requests
pytest
ncclient
paramiko
pyyaml